Extract TimeFormatter from mixin.rb
